Structure fuzzer and properties testing
