+++ rust/src/lib.rs
@@ -1,7 +1,5 @@
-pub mod compat;
 pub mod core;
 pub mod lang;
-pub mod test;
 
 pub use crate::core::*;
 pub use crate::lang::*;
