+++ rust/src/main.rs
@@ -5,8 +5,6 @@
 
 mod core;
 mod lang;
-mod compat;
-mod test;
 
 use crate::core::*;
 use crate::lang::*;
